Save contenteditable content to datebase.
It works with onClick. Click the div and change the content and then hit the save button. If you click on multiple divs it will only save the last edited div. We’ll have to change that. Hope that makes sense. Tested and worked on Chrome and Edge but didn’t work with Firefox.